% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxvolcano.R
\name{coxvolcano}
\alias{coxvolcano}
\title{coxvolcano
perform volcanoplot between beta coefficient and p-values from Cox}
\usage{
library(loopcolcox)

data(cancer)

library(dplyr)

cancer%>%select(3:7)->data

df<-coxbycol(cancer$OS.TIME ,cancer$OS.STATUS ,data)

coxvolcano(df,20)
}
\arguments{
\item{df}{results coxbycol function}

\item{font.size}{size of font in the graph}
}
\description{
coxvolcano
perform volcanoplot between beta coefficient and p-values from Cox
}
\examples{
library(loopcolcox)
data(cancer)
library(dplyr)
cancer\%>\%select(3:7)->data
df<-coxbycol(cancer$OS.TIME ,cancer$OS.STATUS ,data)
coxvolcano(df,20)
}
