% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotnlphr.R
\name{plotnlphr}
\alias{plotnlphr}
\title{plotnlphr
perform negative log10 pval. barplot on Cox results}
\usage{
data(cancer)

library(dplyr)

cancer%>%select(3:7)->data

df<-coxbycol(cancer$OS.TIME ,cancer$OS.STATUS ,data)

plotnlphr(df,5)
}
\arguments{
\item{df}{results coxbycol function}

\item{nb}{number of covariates to plot}

\item{title}{title of the graph}

\item{size}{size of labels}
}
\description{
plotnlphr
perform negative log10 pval. barplot on Cox results
}
\examples{
data(cancer)
library(dplyr)
cancer\%>\%select(3:7)->data
df<-coxbycol(cancer$OS.TIME ,cancer$OS.STATUS ,data)
plotnlphr(df,5)
}
